PlayerController
Work in progress player controller. Loads all our creatures from the game manager into the battlefield. Looks for "friendly" spaces for the creature to move to. We have class-specific functions for switching in and out because each creature can have a "switch in effect" or "switch out effect" that does something while in battle.